requirements.txt
# Core Data Engineering
pandas>=1.5.0
numpy>=1.24.0

# Azure SDK
azure-storage-file-datalake>=12.9.0
azure-identity>=1.12.0

# Databricks
databricks-connect>=13.3.0

# Utilities
python-dotenv>=0.19.0
pyarrow>=12.0.0

# Data generation dependencies
pandas>=1.5.0
numpy>=1.24.0
Faker>=18.0.0
python-dateutil>=2.8.0